[@layer] Change layer order from preorder to postorder
This patch follows from a recent CSSWG resolution [1] that puts the
default layer at the highest priority, and a subsequent spec edit [2]
that changed the layer ordering from the preorder of the layer tree to
the postorder.

Most test changes just reflect this spec change. In some WPT tests,
rule ordering is swapping so that the winning rule will appear first,
so that we can verify layer is working.

external/wpt/css/css-cascade/presentational-hints-cascade.html is moved
from the stable test suite to a new disable-custom-element-default-style
virtual test suite because this is the flag that actually makes the test
result different on test and stable builds. We need to disable this flag
while keeping cascade layers enabled to test the new test case in the
file.

[1] https://github.com/w3c/csswg-drafts/issues/6284#issuecomment-937262197
[2] https://github.com/w3c/csswg-drafts/commit/a6fc16e1b65868c6f984918cc76ad9b238e7139e

Fixed: 1257936
Change-Id: I899de8c8a921b0535296957ee1b0cca1f60a8c61