Split transform-interpolation-computed-value.html
The CSS Typed OM spec and CSS Transforms spec define different
serializations for computed value transform functions and inline styles
respectively. Therefore, transform-interpolation-computed-value.html is
split to appropriately guard the behavior defined in the specifications.

Additionally, inline style serializations for scale are changed in
computed_style_utils.cc when the X and Y scale values are the same, to
match the behavior defined in the CSS transform spec. Values are
omitted if their presence does not change the meaning of the scale().

Bug: 1181086
Change-Id: Iff38bc45075aeed9237739d2979cf12ad5ae2ac7