content-visibility: Skip hit testing children in ng box fragment painter.
This patch ensures that we avoid hit testing children in ng box fragment
painter in the same way we do in the layout object. Since we don't
update other layout phases in display locked content, we can't reliably
hit test there either. Also, we spec the fact that descendants of
display locked elements are not hittestable.

R=mstensho@chromium.org, pdr@chromium.org
Change-Id: I97891536ea9a8beba66116d4b61cf5f1bbfd527f