[FlexNG] Initial set-up for single-line column flex fragmentation
Spec: https://drafts.csswg.org/css-flexbox-1/#pagination

There are four different algorithms for fragmenting flex items
depending on the type of container. The first of which I will be
focusing on is fragmentation for a single-line column flex container
since it will be most similar to normal block fragmentation.

Because we need to know the item location before fragmenting, an
additional layout pass was added to NGFlexLayoutAlgorithm::
GiveLinesAndItemsFinalPositionAndSize() to take fragmentation into
account, at which point we know the final block offset of the item.
The looping mechanism in this method was replaced by a newly added
NGFlexItemIterator, which will iterate over items with break tokens
first, and continue from sibling items thereafter.

The code to set up the constraint space for item layout was moved to
a new BuildSpaceForLayout() method, which is also used in the
final fragmentation layout pass, with an additional call to
SetupSpaceBuilderForFragmentation().

Because we re-layout the entire flex container on every pass, and then
perform a final layout pass taking fragmentation into account, a new
cache slot type of kNone was added to the constraint space to avoid
caching the initial (non-fragmenting) layout pass. This prevents
the fragmented results from getting overridden each time. There may be
potential to avoid this if we can store this information between
fragmentainers, or if we cache it in a new slot type instead of not
caching at all.

With this CL, we can now support fragmentation for a single item in
a single-line column flex container that spans several fragmentainers.
In order to support multiple fragmented sibling items, additional work
will need to be done - including proper block offset calculation,
intrinsic block size calculation, and handling of a break before, all
of which have TODOs.

Some other interesting open questions and areas of investigation
include (TODOs included in the code for these, as well):

- How will the devtools logic work when items have multiple fragments?
- The spec mentions that alignment should happen per fragmentainer.
There will likely need to be work to re-arrange this code in the case of
fragmentation so that it happens after an item has been fragmented.
- An additional layout pass is already made for items that stretch via
ApplyStretchAlignmentToChild(). It would be optimal if we could
make use of this layout pass to apply fragmentation, as well, to avoid
yet another layout pass in this case.
- How to handle flexbox writing mode roots? They are currently treated
as monolithic.

Bug: 660611
Change-Id: I35bcf407bbf741bb5df71f2acc70767be74090c0